+++ src/server/common/Global.ts
@@ -1,4 +1,0 @@
-export interface Global {
-	host: string;
-	port: number;
-}