+++ src/server/domain/EngineConfig.ts
@@ -1,6 +1,4 @@
-import { Global } from "../common/Global";
-
-declare const global: Global;
+import {serverGlobalOption} from "../common/ServerGlobalOption";
 
 export interface EngineConfig {
 	engine_configuration_version: string;
@@ -9,15 +7,17 @@ export interface EngineConfig {
 	asset_base_url?: string;
 }
 
-export const getEngineConfig = (isRaw: boolean): EngineConfig => {
+export const getEngineConfig = (host: string, port: number, isRaw: boolean): EngineConfig => {
+	const hostName = serverGlobalOption.useRequestedHost ? serverGlobalOption.host : host;
+	const portNumber = serverGlobalOption.useRequestedPort ? serverGlobalOption.port : port;
 	const gameContentDir = isRaw ? "raw" : "content";
 	return {
 		engine_configuration_version: "2.3.5",
 		engine_urls: [
-			`http://${global.host}:${global.port}/public/external/engineFilesV1_0_8_Canvas.js`,
-			`http://${global.host}:${global.port}/public/external/playlogClientV3_2_1.js`
+			`http://${hostName}:${portNumber}/public/external/engineFilesV1_0_8_Canvas.js`,
+			`http://${hostName}:${portNumber}/public/external/playlogClientV3_2_1.js`
 		],
-		content_url: `http://${global.host}:${global.port}/${gameContentDir}/game.json`,
-		asset_base_url: `http://${global.host}:${global.port}/${gameContentDir}`
+		content_url: `http://${hostName}:${portNumber}/${gameContentDir}/game.json`,
+		asset_base_url: `http://${hostName}:${portNumber}/${gameContentDir}`
 	};
 };